Redesigned history component to be independent of the GUI
